added .gitignore file
When running scripts in the repository's directory, a "pycache" file is generated. It's pretty much useless junk and can be deleted, but then every time we work in the repository (and run scripts) it will reappear only to need to be deleted again. The ".gitignore" file tells the repository to ignore all pycache files, basically.