Basic Framework Addition (#1)
* Added rough framework

Created a rough framwork for parsing the GenBlock() algorithm into osc messages via pythonosc and sending those to a supercollider script/server. Was able to get the information succesfully into SuperCollider. still using the SCLang interpreter, and looks like that will be the way to go (I know I mentioned earlier that I'd rather just use SCSynth without the interpreter but this will be much smoother, and allow us to easily implement live performance from a midi source later)

* added .gitignore file

When running scripts in the repository's directory, a "pycache" file is generated. It's pretty much useless junk and can be deleted, but then every time we work in the repository (and run scripts) it will reappear only to need to be deleted again. The ".gitignore" file tells the repository to ignore all pycache files, basically.